PI_GCS_2: Patch from PI (Richard Montbrun and Steffen Rau).
we have made a few changes in the source code of our PI_GCS_2 device adapter
fixing some minor bugs:

The first one was from a customer reported in 2011 (and also fixed by us at
this time). There had been a mailing list entry by the customer:

"Together with PI we could improve the communication with our PIZStage (P-721;
Controller E-709). One important step was to connect the PIZStage directly to
the USB port (the E-709 controller has a direct USB connetion and establishes a
virtuell COM port). The baud rate could be changed to 115200. The communication
seems to be more 'direct' in this way. In addition PI provide a new dll for
testing with a small change in device communication. This improves the
communication time as well. PI will probably include this change into the
official dll once it could be confirmed that there will be no other conflict.

The measured times of a z movement of 1 um (as described below) are: 16 msec.”

Another issue concerning the PI C-867.262 controller has been fixed: The
initialization had resulted in an endless loop before (see
int PIController::FindNrJoysticks() ).

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@11951 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd