[nggb] Normalize installation for JARs and scripts.
All JARs now to to subdirectories of $(pkgdatadir). This temporarily
"breaks" the build, in the sense that mmstudio will not be able to find
the JARs in those locations. This should be fixed later so that a
typical Unix-style installation can work (with a correct launcher script
providing the search paths as Java system properties). However, for now,
these changes to the build system are sufficient to ensure that the
files go to a sensible, ImageJ-independent location; the packaging
script for the OS X binary package can deal with placing the files in
the correct final location for the portable app.

Also use semi-correct methods for clean.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@13110 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd