Fix hacks involving CDeviceBase [6/8]: Thorlabs.
Eliminate the fake interface class ThorlabsDeviceBase (from which the
concrete device classes don't actually derive, making the static_cast
from concrete device pointer to ThorlabsDeviceBase* a dangerous hack).

Modify the MotorStage class to use MM::Core* (the call-back-to-core
interface) instead of CDeviceBase methods, so that it works without
having to have a pointer to the CDeviceBase interface of the concrete
device classes (which is what ThorlabsDeviceBase* was used for).

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@13383 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd