Remove BeanShell update() method
It is not clear why this method was ever added, but it reaches into the
private data structures of BeanShell, and will break if we ever move to
BeanShell2 (which uses modern containers for internal data). The
existence of this method also means that we cannot disallow access to
private fields and methods.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@13821 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd