Fix bug causing displayed metadata to occasionally be inconsistent.
I believe there was a race condition here, caused by rapidly changing the
displayed image, so this change postpones the metadata update to 125ms after
the last image update occurs (throwing away any intermediary updates). I was
still able to get inconsistent metadata with this technique, but only once;
the lack of repeatability makes it hard to determine what's going on. In any
event, this is an improvement over the old behavior.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14156 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd