ProblemReport: Make report data persistent
For now, always save to $HOME/MMProblemReport, and delete after closing
the Report Problem window. The idea is to add the ability to find and
reload a leftover (e.g. due to crash) report the next time Micro-Manager
is started.

Data is copied to the MMProblemReport directory as it becomes available.
User-entered data (description, name, etc.) is written with a 1 second
delay to avoid saving the file after every keystroke.

The ability to correctly load a leftover report is implemented in
ProblemReport but not yet tested (and not yet used) as of this commit.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14172 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd