ZeissCAN29: Fix issue causing xy stage to remain busy
Apparently there are cases where an in-flight position of the stage is
reported even after the status has changed to 'settled' (it is not clear
whether this is a bug in some version of the firmware, or expected
behavior). In such a case, a 'target position settled' message follows
the in-flight position message, but there is no 'status change' message
reporting the re-settling. This would lead to the X or Y axis being
considered to be busy again, indefinitely.

This commit is intended to fix this by recording the axis as being
non-busy whenever we receive a 'target position settled' message for
axis devices. It also records that target position. This makes more
logical sense (previously, the target position was parsed but discarded
and the 'target position settled' message was simply ignored for axis
devices); it is likely that the old behavior was simply an oversight.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14322 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd