Start lying to ImageJ about the nature of our image displays.
Make DefaultDisplayWindow descend from JFrame instead of ImageWindow. This
instantly fixes a ton of display bugs and hacks, but of course also wrecks our
interoperability with ImageJ plugins. Hence the addition of DummyDisplayWindow,
which will be the proxy through which we communicate with ImageJ. It's not
entirely clear that this will work well long-term, but the current state of
affairs really can't be allowed to continue -- the amount of technical debt
we accrue automatically by working with ImageJ is simply too large.