Merge DisplayStarter and DefaultDisplayWindow.
Now that DefaultDisplayWindow is no longer an ImageJ ImageWindow, we don't have
to delay its creation as we used to. Instead, we just delay the creation of
all of its UI components (and its DummyImageWindow). Still, this saves us some
conceptual complexity and makes accessing newly-created displays a lot simpler.