Stop MDA during burst without hanging
Prevent acquisition engine from hanging due to full output queue. Also
correctly propagate exceptions and avoid spurious error messages when
terminating with an error or explicitly (via stop()).

This prevents the acquisition engine from aparently locking into
"running" state (as observed via the GUI). This is especially applicable
to out-of-memory situations, in which the output queue is full when
stop() is called to terminate the MDA.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14429 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd