Fix bug causing spurious increase in size of the script area.
This does come at the cost of some additional computation -- now any time text
is added to the script, the entirety of the script contents are examined to
count the number of newlines. But I don't notice any slowdown on my computer,
even for gratuitously large (>1MB) scripts, so I doubt this is a big deal.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14517 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd