Use ByteBuffer/ShortBuffer as the primary backing for pixel data.
The theory here is that we can reap the same performance gains that
TaggedImageStorageRamFast gets. Of course, every time we call getRawPixels()
in the code, now, we have to convert e.g. a ByteBuffer into a byte[], which
has some costs. It remains to be seen how much refinement this approach will
require.