Hopefully speed up config prest switch in main window
The effect of this change is to simply revert one of the changes made in
r12056. This eliminates an unnecessary[1] full system state cache
update, which can block for more than a second on systems with lots of
devices.

[1] It is not _strictly_ unnecessary, in the sense that the display of
other config groups with interacting properties may become outdated
without the cache update. But it does not appear that the change in
r12056 was intentional.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14663 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd