[TIFF saving] Fix splitting by stage position.
There was a bug (and still is, in the master branch) which occurred when
a) split-by-stage-position is turned on, and b) the image tags lack a
position name field. The consequence is/was that the filename generator would
break down, creating multiple writers for a single file, and those writers
would then trample each other, resulting in a corrupted save file.

This change also removes no-longer-needed debugging lines, and removes an
unused parameter from the FileSet constructor.