[Display] Fix a bug causing display scaling to be remembered weirdly.
Specifically, we load contrast settings for a new display *before* we determine
what channel group that display uses -- which means that the display ends
up loading a different set of contrast settings than what it saves, which means
that changes to that display's settings do not effectively propagate to
other displays later.

The fix is to only load contrast settings *after* we have determined the
display's channel group.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14799 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd