[Display] Allow changes in imageCoords to change displayed image.
Previously we were always grabbing the displayed image from the VirtualStack,
which meant that changes to imageCoords could change the scrollbars, but not
the displayed image.

NB previous claims that linking of image coords were working turned out to be
false; image coords were in fact permanently linked due to some aliasing
issues.