[VirtualAcquisitionDisplay] Set the fileName of the ImagePlus correctly.
Previously we were getting it from the AcquisitionVirtualStack, which was
always returning null. Now we get it from the ImageCache, which gets it from
the Storage. Consequently, TaggedImageStorageRamFast had to be updated so that
it was capable of remembering that it represents disk-backed storage, instead
of always returning null in getDiskLocation().

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@14905 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd