[Display] Don't permanently steal ImageJ's menubar.
For some reason, when we set our display windows to use ImageJ's menubar, this
actually "steals" the menubar from ImageJ; repeated stealing can cause the
menubar to be lost altogether. This change causes us to give the menubar back
when our menus are deactivated, which should fix the issue...though it's still
super-worrying that the menubar can apparently only be in one place at a time.
Also, "menu" no longer looks like a word.