[Contrast links] Ensure unique IDs even when channel names aren't available.
You used to be able to get two linkers with an ID ending in "_null" where the
channel name should be. Now we catch that and use the channel index instead.
Which isn't exactly ideal, but at least now if you don't have channel names
set, linkers for the same display on different channels won't link to each
other.