[Fullscreen mode] Completely rework our approach.
Now instead of creating a duplicate DisplayWindow that operates in fullscreen
mode, we create a new JFrame and transfer the contents of the "source"
DisplayWindow to it. This preserves the state of the DisplayWindow (e.g.
opened tabs, ROIs drawn on canvas), keeps any custom controls visible, means we
don't waste resources on calculating histograms for a hidden display, keeps
the ImageJ "Windows" menu accurate, and so on.

Probably though there are some bugs introduced by the new approach, but they're
unlikely to be worse than what we had before.