[Image display] Fix a bug causing invalid image coordinates.
We were failing to recognize properly when time/Z/channel axes were not present
in the Datastore, and consequently trying to look for images at locations where
they did not exist. This fixes the SplitView plugin and any other situation in
which someone might try to set "minimal" image coordinates that don't have all
of the time/Z/channel/stage position axes.