[Display] Fix error spew.
We have a hack in our drawing code (when setting LUTs in ChannelControlPanel)
to set the ImageJ CompositeImage's "currentChannel" property to -1. This fixes
an issue in Color and Grayscale drawing modes where the LUTs would be wrong,
causing incorrect colors/scaling. However, in Composite mode it could cause
errors when drawing images. These errors were (apparently) harmless in MM2.0,
but they happened frequently. Now we only do the ImageJ hack when we're
actually in Color or Grayscale drawing mode, which seems to fix the spew
without breaking the display in any way.