[Album] Make album its own top-level API entity.
This splits it out of the Data API, which is good because this way the Data
API doesn't depend on the Display API. In the process I also removed some
obsolete methods from CompatibilityInterface.

NB in this commit the coordinates of images in the Album are off. I'll
fix that next.