[Various] Fix memory leak and remove obsolete API methods.
This fixes a memory leak that was causing the DisplayWindow (and thus, the
associated image data) to be kept in memory after it was closed. Mostly this
involves properly unsubscribing from events, destroying the old
CanvasPaintPending global singleton, and fixing a bug in closing the
DummyDisplayWindow.

Along the way, the AcquisitionManager's record of extant acquisitions was
nuked, causing many API methods to become invalid; those methods were therefore
deleted. The few remaining methods of AcquisitionManager are all passthroughs
now.

Also, some compilation warnings in the MDA dialog have been removed (albeit,
not all of them).