[Display colors] Don't record colors in default display settings.
They're automatically recorded by the histograms, and we want those values to
take precedence over other values that are after all just a different set of
profile-stored values. No need to have the profile fighting itself for
precedence, and this way the user doesn't have to manually save their colors.