[Link buttons] Fix link button display.
The DisplayGroupManager was finding out about new displays too late because of
changes we'd made to the order in which events get fired. Now we post both a
NewDisplayEvent (shortly after the DisplayWindow object starts existing) and
a DisplayAboutToShowEvent (right after all of the GUI objects are created),
so objects that care about one or the other can get them when appropriate.

Also adds the DisplayWindow to the values accessible in NewImagePlusEvent;
this is a prelude to some work on fixing the PatternOverlay plugin (unrelated
to this commit otherwise).