[PatternOverlay] Update to new event types.
This is still buggy for multi-channel setups: the user snaps, the first image
is added to the Snap/Live view, the plugin is activated, it adds an overlay
to the (single-channel) ImagePlus, the second image arrives and the display
shifts to a multi-channel CompositeImage, and the overlay is lost.

In the future ideally PatternOverlay will use MM's explicit overlay system
instead of ImageJ's one, which will render that bug irrelevant.