[Display scrollbars] Fix EDT hang.
The textfields we'd been using for displaying the current image coordinate
weren't threadsafe, leading to a deadlock situation between the thread changing
the text and the thread trying to draw the textfield. Now we use buttons
instead, and when you click the button you get a separate textfield to type the
new coordinate into.

Along the way, the initialization of the DefaultDisplayWindow's GUI objects
was moved to the EDT.