[Link buttons] Overhaul implementation and improve UI.
You now get a popup menu every time you click the link icon, with a list of
displays to link to. The implementation of the SettingsLinker class has been
simplified as the lists of linker groups and which linkers are active now are
handled by static singletons, instead of each SettingsLinker keeping track of
its relationships and having to handle ensuring those relationships were
reciprocal.