Histograms: improve link buttons
New inspector windows should open with full knowledge of all extant displays'
histograms. Histograms and other InspectorPanels clean up properly when
inspector frames are closed.

Removed the Histograms interface class, which was wholly useless (only had one
implementation, and the only code that ever spoke to that implementation didn't
use the interface).

Unfortunately still buggy, in that we now have multiple SettingsLinkers for
each display when there are multiple InspectorFrames open, which leads to
redundant entries in the link button dropdown menu. Looks like we'll be moving
to display-specific ContrastLinkers in the future...