[Canvas update thread] Simplify main loop.
This may also fix an intermittant NullPointerException when closing windows;
it's hard to know without a lot more investigation though.

Added a cap to the number of times we wait for the canvas paint pending to
clear before we force it clear and continue on, which should hopefully serve
as a failsafe for "display break" scenarios (though if the canvas update
thread exits unexpectedly, then that will also "break" the display).