[DisplaySettings] Remove imageCoords property.
This requires rewriting the ImageCoordsLinker and part of ScrollerPanel. In the
process I added some optimization to ScrollerPanel to prevent it from making
too many GUI updates (a similar concept to the way CanvasUpdateThread won't
let lots of image draw requests pile up). The overall goal here was to reduce
the number of NewDisplaySettingsEvents that we generate and therefore have to
process; the previous method generated at least one such event for every
scrolling action or new image, which is potentially a lot of extra work to do.