[Metadata] Auto-populate userData property.
Any values that aren't a) part of the base Metadata, or b) part of the
scopeData property, will be shoved into userData.

This also changes MDUtils.getKeys() to only return keys that point to values
that are not null. This protects us from JSONExceptions when we follow up with
attempting to get() the value pointed to by the key, and I believe the only
places where we use getKeys() are places where we're going to follow up with
a get() call.