[Multipage TIFF] Fix loading of 2.0 files.
We'd been trying to load all of the "scopeData" properties into the "userData"
section, which was failing hideously because the properties were dicts instead
of something that can go into a PropertyMap. Now we correctly ignore all
scopeData properties when loading userData, and messages when an unexpected
type appears are debug-level instead of error-level.