[OME metadata] Fix saving OME Metadata.
This is ugly. Our OME metadata writer was bombing out if we tried to save
information for sites "out of order" -- that is, saving metadata for
stage position index 1 before index 0, for example. The OME metadata code is
basically a complete unknown, and our library is out of date besides. So the
workaround for now is to sort images by stage position index when saving, so
we always save them "in order". But we have to do this in the Datastore because
that's the only place that actually knows what stage position indices are
available.

Like I said, ugly.