[MDA dialog] Prevent duplicate channel entries.
When adding channels, the dropdown menu to select which channel to use will
only show channels that are not already part of the acquisition. Attempting to
run an MDA with duplicate channel entries would create an error, so such
setups were invalid already; the GUI just didn't prevent you from making them.