[Display] Fix Windows layout/display issues.
This prevents the display window from extending undernearth the
taskbar, and also appears to fix a flickering issue we were
having when the window is wider than it "needs" to be, though
I'm not certain what exactly fixed that.

Still remaining: display issues when in fullscreen mode and
zooming out.