[Display] Fix layout issues.
We now no longer override the pack() method to ensure the canvas' size is
correct, as the canvas asks the display window how much space is available
for it instead of trying to derive its own size. Some adjustments to other
parts of the code had to be made due to this change, but it's a major net
positive as the amount of "manual layout management" we're doing is much
smaller now.