[Loading] Fix importing 1.4x-era scope metadata.
Now we check the metadataVersion field, and only do the "import all remaining
fields into userData" action if the metadata version is 10 or older. This fixes
issues where we try and fail to import data that the
MDUtils.extractUserData() method can't parse, and also keeps us from
overwriting the actual userData structure when a more modern file is imported.