[Version check] Add an ImageJ version check dialog.
This shows a warning if the version of ImageJ the user is using is not known
to be compatible with Micro-Manager. Important as users may use Fiji with any
number of ImageJ versions and many of them are very sub-optimal (e.g. 1.49u
causes the MDA image display window to be eternally blank...).