[Movie export] First working commit.
This generates a stream of numbered PNGs in the current working directory.
To get the image export working, I had to modify MMImageCanvas to do manual
double-buffering, which isn't an awful idea anyway. But it also gives us
access to the drawn image in an intermediate state (i.e. not sent directly
to the graphics card for display) that we can then save to disk.