[MDA dialog] Add sanity check for long exposure times.
If the user has set up an exposure more than 30s long (using the channel
table, not by copying from the main dialog), then they'll get a prompt when
they try to run the MDA asking if they're really certain. Added after a typo
during a test resulted in a 5-minute-long exposure...which the camera was
happy to do but which rendered the program unusable while the exposure was
running.

The dialog has a "do not ask me again" prompt.