[Display] Replace CanvasUpdateThread with non-threaded approach.
The new CanvasUpdateQueue draws the canvas by putting code on the EDT to
perform draw updates, and cycles itself by listening for the
CanvasDrawCompleteEvent to indicate when the canvas is free to draw again. This
is cleaner, removes some hairy wait logic (including all use of the "canvas
paint pending" concept), and of course removes a thread.

The animation system is similarly reactive, listening for
CanvasDrawCompleteEvents and scheduling new animation updates when those
events are received, instead of blindly requesting the display update at set
intervals. This fixes an issue where animations were "stuttery", hopping back
and forth and only stochastically making progress, due to race conditions and
the hard-to-predict nature of which images in the animation actually get drawn.