[Multipage TIFF] Fix error causing truncation of OME metadata.
More generally, any descriptor tag written with
MultipageTiffWriter.writeImageDescription() could potentially have been
truncated, as it was reserving String.length() bytes, but actually writing
bytes based on converting the String to a ByteBuffer, which could be longer
(for unexamined reasons). Now we reserve space based on the size of the
ByteBuffer instead.