[Display] Remove redundant redraw requests.
Our histograms have logic that: if they are the first channel, a new image
arrives for something that is *not* the first channel, and their color is
white, they reload the display settings and force a redraw. Unfortunately, if
their color is actually supposed to be white (as opposed to temporarily hacked
to white because they were for a singlechannel dataset before the new image
arrived), then this happens on every single new image, making the display badly
nonresponsive. Now we only reload display settings if we know the color is
wrong. Ideally in the future we won't need the color hack either.