[DisplayManager] Fix issue with managed datastores.
The DisplayManager didn't know about DisplayWindows that ImageJ didn't know
about, which meant that if you a) created a datastore, b) created a display for
it, c) managed the datastore, and then d) added the first image to the
datastore, it was effectively unmanaged because the DisplayManager didn't
know about the DisplayWindow (which "didn't exist" at the time it started
managing the datastore) and thus didn't listen for its request-to-close event.
Now DisplayManager knows about *all* DisplayWindows, so this problem is
obviated.