CSU: Avoid using state interface for speed (Pariksheet Nanda)
[See http://micro-manager.3463995.n2.nabble.com/patch-CSU-DriveSpeed-property-stopping-HCW-editing-tp7585634.html]

To address the problem discussed in the earlier thread [1] I've
reassigned the DriveSpeed property of the Yokogawa CSUX, CSU22 and
CSUW1 to be a GenericDevice instead of a StateDevice.

I suspect the StateDevice was problematic because the there are up to
8500 speed states for the DiskSpeed (1,500 to 10,000 on our 10,000 rpm
CSUX scan head here).  So I don't think DriveSpeed is a good fit for
the StateDevice API.

With DriveSpeed as a StateDevice, editing the Micro-Manager
configuration would often fail in 4 or less tries on our system here.
After changing DriveSpeed to be a GenericDevice, cycling Micro-Manager
produced no problems with the CSUX even after 180 cycles of editing
the configuration.

[1] http://micro-manager.3463995.n2.nabble.com/CSU-22-settings-td7582455.html

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@15674 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd