Live: be lessing willing to reset the display.
In particular, we no longer reset when the core camera changes or when the
channel name changes. The latter was causing endless resets in multicamera
mode, and the former was causing endless resets when the Split View plugin is
used. As a side-effect, if you go from multicam to single camera, the live view
will have a "stale" channel (that never receives new images) until you close
and re-open it.