OughtaFocus: Add methods MedianEdges and FFTBandpass (Jon Daniels)
From Jon:

It adds two focus scoring algorithms to OughtaFocus, first a slightly
modified version of the algorithm in the JAF(H&P) autofocus device
described on the wiki.  Second and more notably, an FFT bandpass
algorithm developed by Bill Mohler for their diSPIM data.  The focus
score consists of the average FFT intensity in a certain range of
spatial frequencies.  Because it uses a modified FFT (without
autoscaling) there is an entire static nested class to compute the
unscaled FFT based on ImageJ code with Bill's tweaks included.  The
upper and lower cutoff frequencies are represented as percentages set in
properties that I added to the OughtaFocus device; these properties are
only used by this algorithm and provide much-needed flexibility to set
the desired range of frequencies (the default values are set to the ones
Bill likes for their datasets).

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@15751 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd