Display: handle color mode/LUTs correctly with no inspector window.
Previously, if the inspector window was closed, then LUTs and custom channels
could not be set; we were stuck with default ImageJ colors. The addition of a
new "LUTMaster" class that is responsible for adjusting LUTs, which is used
by the GUI controls, fixes this. Well, more specifically, the fact that the
DisplayWindow can tell the LUTMaster to update it fixes the bug; the LUTMaster
just neatly refactors things and separates the controls a bit from the actions
those controls represent.