Display: fix a bug in numbering of anonymous, frozen datasets.
The "Untitled Data #N" would keep incrementing every time the title was
re-calculated, because the generated name couldn't be stored in the dataset, as
it was frozen. So now we keep an external mapping of display hashcodes to the
generated names we make for those displays. I don't really see that we have any
other option.