MDA: use ChannelSettings to load/set channel colors.
Thus they make the transition to the DisplayWindow again (after the recent
change to the DisplayWindow's initial display settings), because they update
the profile with the new preferred color for that channel. And changes to the
channel color after the MDA is complete will be reflected in future runs of the
MDA, once the MDA dialog is refreshed, anyway.