Display: fix some instances of the "grayscale bug".
We now force-reapply display LUTs in a few areas. The root cause of the
grayscale bug is unknown, and at the moment I assume it's some weird drawing
bug, possibly in ImageJ.

This is not a complete fix for issue #49 as we still switch from color to
composite mode when duplicating a display.