Incomplete work on exposing a more general "DataViewer" interface in the API.
The chief problem is that this doesn't let you use histograms, which are heavily tied into
how the existing DisplayWindow class works -- and histograms are the #1 reason why you would
want to generalize the existing display system (so you can have alternate displays that
still use the inspector). So that's not very helpful.