Histograms: completely refactor when/how histograms are calculated and displayed.
The goal here is to make it so that the HistogramsPanel can be used for DataViewers instead
of just DisplayWindows. This entails the following changes:

 * Remove ChannelHistogramModel; its former users now do what it did directly, except for
   histogram calculation, which was moved to CanvasUpdateQueue.
 * Thus, ChannelControlPanel now directly updates the DisplaySettings when it is manipulated
 * Add events to allow ChannelControlPanel to request and receive new histogram information.
 * Add ContrastCalculator class for generating histograms, and HistogramData class for
   encapsulating that information.

The practical upshot of this is that we now calculate new histogram information immediately
prior to drawing images (which should remove some redundant draw requests). But we still
assuredly have plenty more bugs to sort out...