MainFrame: fix window sizing bug.
If the default profile did not exist, and the user cancelled the intro dialog rather than
complete logging in, then the main window would save invalid sizing information to the
profile (because the main window had not been displayed yet and had a component with a
preferred size of 9999x9999). This then prevented the program from successfully creating the
main window in future runs. The fix is to avoid creating such an absurdly-sized component in
the first place.