Display: give the first new image window focus, instead of the inspector window.
The first new image window created in a session in turn creates a new inspector window, which
had been stealing focus; now the window recovers that focus. This fixes a minor issue with
display of the histogram controls in the inspector not being updated until the window was
manually clicked on.