Display zoom: use ImageJ-style zoom selection instead of powers of 2.
This means that if the user uses the ImageJ magnifying glass to zoom and gets to a zoom level
like 75%, then the zoom buttons will be able to get them "back" to 100% zoom; under the old
behavior this was impossible unless they went back to the magnifying glass tool.

This fixes issue #114.