DisplayManager: rework how getCurrentWindow() operates.
It used to rely on ImageJ's WindowManager class, and thus would return null if
the topmost window was closed. DisplayManager now maintains its own internal
history of which windows have been raised to top, so if a window is closed,
then the next-topmost window should be automatically made available by this
method.