Shutter: add ShutterManager to handle GUI synchronizing with Core state.
This eliminates a bunch of tedious bookkeeping logic between Studio and
MainFrame to ensure that MainFrame's GUI is manually kept up-to-date whenever
someone calls the shutter methods in Studio. Instead, ShutterManager methods
are used to update the shutter state, and it posts events that keep the GUI
up to date.

This also adds a preliminary Quick Access tool for changing shutter state.
Currently it only turns autoshutter on and off.