MetadataPanel: fix issue #135 (metadata display blanks out).
The root cause was us doing repeated display updates of the same image, which
meant that every metadata key was "unchanging" and thus should not be
displayed. We now refuse to redisplay metadata for the same image (tracked by
its UUID) twice in a row. While in the area, I also changed our timer-based
system to a queue-and-thread-based system, which should be less prone to
potential race conditions.