DisplaySettings: allow DisplayWindows to segregate their settings.
That is, not all DisplayWindows need necessarily store their preferred
DisplaySettings in the same part of the profile; the [get|set]StandardSettings
methods now accept a string key to separate them. This is used to allow the
Snap/Live view to store its settings independently of the MDA view(s).