Display: fix issue #139 (memory leak from creating/destroying displays).
The root cause was that contrast linkers were not being destroyed when a
display is destroyed, due to a bookkeeping error. This, combined with the fact
that all SettingsLinkers of any type are stored in a static mapping, meant that
references to the destroyed displays were being kept around, leaking memory.