Acquisition: add runAcquisitionWithSettings() method.
The problem with the existing runAcquisition() methods is that they always
overrode any SequenceSettings as set by the setAcquisitionSettings() method
with the settings currently in the MDA dialog. This method bypasses the MDA
dialog altogether in favor of using the provided SequenceSettings, and should
be the preferred way to manually configure an acquisition.