Snap/Live buttons: fix error causing them to redundantly unregister from events
The HierarchyListener we were using fires more often than I'd anticipated,
causing the buttons to unsubscribe early (which broke disabling of the Snap
button when Live mode turns on/off, for example) and then trying to unsubscribe
again when the Live window was destroyed, which caused an
IllegalArgumentException. This is now fixed.