Inspector: fix issue #154 (inspector frame too small).
The cause of this was that the individual histogram panels
(ChannelControlPanels) were only initializing their UIs after having received
histogram data, which happened some time after they were added to the
HistogramsPanel and the Inspector layout was refreshed. Now
ChannelControlPanels can tell the HistogramsPanel to update its layout
themselves, propagating that change up to the inspector.