Canvas: remove manual double-buffering; tweak ExportMovieDlg to adjust.
Instead of double-buffering every single canvas paint, we now provide an
external method to paint to a provided Graphics object, which the canvas'
normal paint method makes use of. ExportMovieDlg can then ask the canvas to
paint to its own Graphics derived from a BufferedImage, but we don't have to
pay the performance penalty of drawing to a BufferedImage and then to the
screen for every single paint the canvas does.