DisplayManager: tweak order of events slightly.
The previous order meant that the Inspector was made aware of new
DisplayWindows prior to those windows having set up their UIs, which in turn
meant that when the inspector tried to ask the DisplayWindow for histogram
data via a HistogramRequestEvent, the CanvasUpdateQueue wasn't yet subscribed
for events and thus didn't notice.

Arguably we should go with a more structured system here instead of relying
on order-of-events; this is just a patch job.