Storage: add explicit putImage() and freeze() methods.
Previously the Storage had been relying on NewImageEvents and
DatastoreFrozenEvents to know that it had to perform these actions, which a)
was loose coupling of something that should be tightly coupled, and b) meant
that any exceptions that occurred when performing those actions would be
swallowed by the EventBus. Which is bad.