Metadata: remove gridColumn and gridRow properties.
These appear to have been confused with the MultiStagePosition's row/column
properties, which can be accessed elsewhere; the gridColumn and gridRow
properties of the Metadata object were never set anywhere and thus always null.
(MMAcquisition was setting a GridColumn and GridRow in the SummaryMetadata's
JSON representation, but that likewise couldn't possibly accomplish anything).