Snap/Live: fix a bug causing slow display updates after changing the channel.
Our snap/live display tries to dynamically adjust the rate at which we feed
images into the Datastore to account for the amount of time it takes to
process and display those images. However, we weren't reseting our time-
tracking when the display gets reset (as it is whenever the channel changes),
which caused us to perceive an unusually long time passing between display
updates, causing us to back off the rate at which we fed images in.