DefaultDisplayWindow: don't block calling thread when creating GUI.
Previously when we ran makeGUI(), we would block until it finished, because we
needed the display to be fully set-up before we could go on to display images.
makeGUI() has been modified to handle that display logic itself, so the calling
thread no longer needs to block.

This also modifies the getIsClosed() method to return haveClosed_ instead of
examining the current state of the window.