HistogramsPanel: fix a bug relating to empty Datastores.
The bug was that the HistogramsPanel would assume the Datastore had no
channel axis, and would create an invisible ChannelControlPanel for the
nonexistent channel axis. When images with an actual channel position of 0 were
added to the Datastore, no corresponding ChannelControlPanel would be created
as the HistogramsPanel thought it already had all the panels it needed.