Album: set SummaryMetadata, including channel names.
The album uses the current channel in the current channel group as the channel
name, and will create a new album if this value changes. This is a change in
the API, but I believe it's the right behavior considering the expected use
cases for the Album. Proper channel names mean that the image color and
contrast settings will automatically be correct and linkable across displays.