Pipeline: add isHalted() method.
This also changes the semantics of AcquisitionManager.isAcquisitionRunning()
so that it returns true if the Pipeline is not halted. In turn that means that
AcquisitionManager.runAcquisition() will not return until after all images from
the acquisition have entered the Datastore. This fixes a potential
bug where code that waits for an acquisition to finish (i.e. the acquisition
engine to finish executing) could fail to have access to the complete dataset,
as the pipeline was still processing images.