MMStudio.setROI: add support for multi-ROI setting.
We decompose the composite ROI into a set of bounding boxes, so e.g. using an
ellipsoidal ROI in the composite ROI will still get you a rectangle. It'd be
nice to error out instead in this situation, but that would require being able
to detect non-rectangular components of composite ROIs, which appears to be
nontrivial.