Display: don't redundantly redraw when autostretch is enabled (#278).
Autostretch is performed immediately before drawing the new image, but it also
updates the DisplaySettings of the display, and any time that happens, the
display redraws itself. Thus we were trying to draw each image twice whenever
autostretch changed the contrast settings. This change adds a new
setDisplaySettings method to DefaultDisplayWindow that allows disabling of the
redraw.