MultipageTiffReader: don't create Comments annotations uselessly.
It turns out that acquisitions start with a blank string as a comment for every
image, which meant that on re-opening the file, the Reader would try to convert
the "comments" into an annotation even though there weren't any actually valid
comments. This is basically harmless, though; the original files were not
modified in any case.