SnapLiveManager: fix a deadlock.
If we try to stop live mode when we're in the middle of calling
startContinuousSequenceAcquisition (possible due to core callbacks), then we
can deadlock because the grabber thread is waiting for the core camera lock,
which the main thread has, and the main thread is waiting for the grabber
thread to exit. This fixes the deadlock by recognizing this situation and
refusing to start/stop live mode when in the middle of calling
startContinuousSequenceAcquisition.