MMStudio: fix a bug in setExposure.
The symptom is that when running Live mode, it was impossible to change the
exposure time via the text field in the main window, because it would be
immediately reset to the old value. The cause was that a GUI reset was
triggered by changing the exposure time, which resulted in the GUI verifying
its exposure time by pulling the old value out of the profile.