SnapLiveManager: fix channel names when snapping images.
Earlier changes to limit the number of calls into the Core had caused us to
operate with an invalid count for the number of core camera channels, which
lead to us not generating proper channel names, which lead to an unnecessary
display reset if the user launched MM, snapped an image, and then started
live mode (since live mode would cause us to grab the number of camera
channels, thus noticing the channel name discrepancy).

Anyway, this fixes issue #356, as well as the display colors being wrong
for snapped images when live has not yet been run in this session.