Plugins: add DisplayGearMenuPlugin type.
This plugin type is used for entries in the gear menu in image display windows.
It is structurally very similar to the MenuPlugin type, but appears in a
different part of the GUI and receives the relevant DisplayWindow in its
onPluginSelected() method.

This also moves SortedMenu out of DefaultPluginManager, and adds a
corresponding SortedPopupMenu class for JPopupMenus.