Menus: add menus to all non-image windows (issue #183).
This modifies MMFrame to create a JMenuBar that will be shown when the frame
has focus. This required reworking how we create our menubar since we now need
to be able to create multiple copies of it. MMFrame exposes a constructor that
provides the old behavior, which is necessary for DisplayWindows as they want
to show the ImageJ menubar instead.