Alerts: add AlertsWindow; tweak Alerts API.
The AlertsWindow holds all Alerts, rather than their being shown as undecorated
dialogs. Added a button to the main window that indicates if there are alerts
and opens the AlertsWindow when clicked. There's still substantial GUI work to
be done here.

Some of the API methods now take a "title" String parameter.