StorageSinglePlaneTiffSeries: throw exception if image lacks PositionName.
This fixes issue #378 in a somewhat lazy fashion.

The change to MDUtils removes an implicit position name derived from the
(bogus) "PositionIndex" JSON tag that was present in TaggedImages output by the
Core. MDUtils.getPositionName() is only invoked by DefaultMetadata and the
multipage TIFF code; the latter already checks for a proper position name and
is thus unaffected by this change, the former will now not report a position
name if there is none in the source tags.