mmstudio and plugins: Windows now listen for the ShutdownCommencing Event.
By listening to this event, these windows will close when shutdown starts and do not rely on a System.ext(0)
to close.  The plugins that I changed (not all yet), listen to the external event.  It may be good to add this function
to the MMPlugin interface so that plugin writers realize that the may want to implements this.